Вынес структуры моделей в отдельный файл
