+++ ImageFeed/Profile/ProfileService.swift
@@ -1,29 +1,9 @@
 import UIKit
 
-struct ProfileResult: Codable {
-    let userName: String?
-    let firstName: String?
-    let lastName: String?
-    let bio: String?
-    
-    private enum CodingKeys: String, CodingKey {
-        case userName = "username"
-        case firstName = "first_name"
-        case lastName = "last_name"
-        case bio = "bio"
-    }
-}
-
-struct Profile {
-    let userName: String?
-    let fullName: String?
-    let loginName: String?
-    let bio: String?
-}
-
 final class ProfileService {
     
 //  MARK: Объявление строгого singleton:
+    
     static let shared = ProfileService()
     private init() {}
